refactor(api): move avatars to modules/ and put its work in the service
